docs: add offline section to the README
Closes #95
